+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 2.0.0
+- Update all weights to be in line with Apotheosis weight changes
 - `school_attribute`, `spell_effect`, `mana_cost`, `spell_level`, and `spell_trigger` now support optional filtering and multiple schools.
 - Updated Chinese Translation (`zh_cn`) - Thanks ZetaY!
 - Updated other affix translations (may interfere with the updated Chinese and other keys)